Chore: auth type => type 폴더로 이동
