+++ src/types/auth.ts
@@ -1,0 +1,6 @@
+export interface AuthValues {
+  email: string;
+  name?: string;
+  password: string;
+  passwordConfirm?: string;
+}
